[SATest] Fixed bug in SATest with writing reference and NEAT data to a file with empty name.
CR: http://cvcc-ubu-01.iil.intel.com/cr/r/5047/
Tests: compiler pre-commit.

The problem was with the exception type thrown by ReferenceRunner.
IOError exception is supposed to be used if there is no file on the disk
with given name, but we used it also if the filename is empty.
Fixed the exception type and sorted headers in the source files.