CSSD100019990: do not vectorize 'unreachable'
Kernel provided by user contained invalid code (dereferencing an
uninitialized pointer) which LLVM identifies and replaces with:
    tail call void @llvm.trap()
    unreachable
This fails an assertion in the Predicator which expects all BBs to terminate
with branches or returns.
This fix will prevent vectorization if 'unreachable' instructions are
found in the function. Since this is done after dead code elimination,
finding an 'unreachable' instruction implies esoteric code we would
not want to apply the vectorizer to.
A counter was added to allow tracking of functions not being vectorized
for this reason.

PC: http://ocl-jenkins.iil.intel.com:8080/job/PCompiler_grapapor/11/
CR: http://cvcc-ubu-01.iil.intel.com/cr/r/5018/ by Aleksey, Amjad