Implemented OpenCL address-space Alias Analysis
This commit provides an immutable pass for the anti-aliasing
information available through the OpenCL address-space of values:
(1) alias(Loc& A, Loc& B) returns NoAlias for A, B of different address-spaces
(2) pointsToConstantMemory(Loc& L) returns true for L of OCLAddressSpace::Constant

alloca instructions may be used for Private and Local allocations, but the alloca's
themselves do not have a an explicit addrspace(). We therefore follow the users and
base addresses of pointer values that do not have an explicit address space until
we find one (if we cannot, we consider the address space to be Private).
We behave conservatively (i.e. don't know) for addresses which escape the analysis
via 'ptrtoint' / 'inttoptr': we avoid declaring such addresses as Private even
if we find no explicit relation to any other namespace since the address may be
still related to one if the int value is ever casted back to a pointer (note that
we don't care about the pointer escaping to int if we do find a relation to
a non-default address space).

PC: http://ocl-jenkins.iil.intel.com:8080/job/PCompilerBuild_grapapor/16/
CR: http://cvcc-ubu-01.iil.intel.com/cr/r/5033/ (Aleksey, Amjad)