Fix for CSSD100020372: A sync bug in all CL-GL objects that are created with CL_MEM_WRITE_ONLY flag
Problem occurred for GL surfaces that are shared with OpenCL with CL_MEM_WRITE_ONLY flag, when the GL surface is synced back to GL (using clReleaseGLObject) the data transferred from OCL to GL is wrong.
Bug was that OCL runtime wasn't reading the texture data in acquire command for objects with CL_MEM_WRITE_ONLY, I guess it was some kind of optimization, but this is a wrong optimization !
Even with CL_MEM_WRITE_ONLY, OCL runtime is not allowed to wipe the GL surface data, because user can write to only a small region in the surface through OpenCL or even not write to the surface at all.

+ Removed exporting the CL-GL sharing extension APIs when it's not supported (related to CSSD100020286)

Reviewed by Aharon.

PreCommit:
http://compute-ci-jenkins.iil.intel.com:8080/job/PRuntime_nmeraey/1/