Vectorizer bug fixes
- Fixed bugs with builtin functions (SCALAR should be  1, not 0; Packetizer
now asks for 'vector-width' version on the function (instead of log()).
- Improved builtin functions (A function attribute is now attached to
'vectorizer builtin' functions (i.e. functions formerly defined in the
'builtins' module). Note that this is only a subset of the stubs generated
by the vectorizer.
- Hardened argument scalar-to-vector expansion: using a broadcast for random
arguments is somewhat risky (some pass might notice we're extracting those
scalar values and replace the fake vector with the original scalar argument).
Code now emits a function call that produces an opaque vector from the scalar
argument.
- Support multi-scalar values for arguments: Arguments now produce correct
multi-scalar values, based on a fake wide value which may be a vector or an
array depending whether the argument is primitive or an aggregate. If the
argument's type is packetizable we use the vector value from the VCM (and make
sure it exists). Unlike for instructions, vector and multi-scalar values for
arguments placed at the basic block where they are first used will not
necessarily dominate all their uses (since the argument itself dominates the
entire function). We therefore hoist the vector and multi-values definitions to
the nearest common dominator of their current basic block and the basic block
of each new user. This adds TreeDominator as a required analysis for the
Packetizer.

Old/split llvm commit ID: a859136da09dcc2028d2e1f0e66ab5be8d93085f