Fixed build on Ubuntu64.
This warning help me to realize that previous fix of Android64 issues
was not a workaround of Android NDK bug. We had out of boundary access
in the loop, which is undefined behaviour by language specification.
Thanks to GCC warning on Ubuntu that failed our build because it was
treated as an error I noticed that bug in our code.