[LLVM 3.6 UPGRADE] Upgrade LLVM IR based tests
This is an agregate of impact of LLVM IR evolution on OCL BE tests based
on LLVM IR. Most of changes are expected to be done in the LIT tests.
Suggested approach is to use an appropriate version of Clang to generate
updated LLVM IR including the metadata and the debug intrinsics.
This commit contains an exmple of such test generated with LLVM and Clang
built from merged 'origin/intel/features/cpp and intel/upgrade_36'
branches. This merge can be found on private/aelizuno/upgrade_36_and_cpp
branch.

TODO: Add more MetaDataApi tests to cover the Mako schema completely.