[LLVM 3.6 UPGRADE] Use legacy DataLayoutPass
DataLayout pass has been renamed to DataLayoutPass. It looks like this
has broader impact than that. Later I've got an impression what it is not
needed anymore and llvm::Module::getDataLayout should be used instead.

FIXME:
So during LLVM 3.6 upgrade the pass was sporadically replaced with the
new approach and with legacy DataLayoutPass approach. It shouldn't be
difficult to track usage of DataLayout and use one approach everywhere
in OCL BE instead of this mixed style.