Fixing assert in demangler in WI analysis (CSSD100021759). Add regression LIT test for demangler assertion fix.
When a scalar select is used in a devergent section
OCLBuiltinPreVectorizationPass creates a fake for it, Predicator
masks this faked function. Then WI analysis expects fake builtin
prefix _f_v. to be at start of the name, but it is not, since the
name was mangled by Predicator later in Vectorizer:

%2 = call float @maskedf_0__f_v._Z6selectffi(i1 %Mneg, float undef, float
%%scalar, i32 0)

The patch makes WIAnalysis to demangle the name first and then get
scalar function name.