Refactored the way STI types are created.
The main motivation for this refactor is to solve the possible loop in a better way than it was done so far.
Processing debug info could get into a loop because of records (structure, class, union).

In the old implementation a loop was recognized in the createType function, but were only handled when getting to the record type, then a forward declaration was created instead of record definition, forcing the loop to break.
That means, some types may end up having two or more STI entries, which will be deleted in createType function to prevent duplication.

The new implementation makes the createType much simpler, all it does is to check if the type is already in the map, if exists just return it, otherwise call the right function to create it.
In order to break the loop, the STI entry must be created and added to the map first thing when creating any type. This way if there is a loop, when getting to the createType funtion it will find that entry and returns it, the loop is broken.

In addition, when calling the record type creation function for a named record, it will implicitly create a forward declaration record and map the defined record to it.
During STI entry creation all entries will point to the forward declaration record, even the S_UDT entry.
Later, before emitting the STI entries, those S_UDT entries will be fixed to point record definition.


Old/split llvm commit ID: 836350632afa6fa77b405d5c4f5cf7829543abc0