Fix for CORC-902 [BXT] Half 2.0 subtests vstorea_half_rtn/vstore_half_rtn fails on Win10 BXT-P with error 'Failure at [0] with 0x1.000000p-24: *0x0001 vs 0x0000'.
The bug affect ISA <=AVX. AXV2 provides an instruction for such
conversion.

Validation fails when the rounding algorithm faces a boundary
float value: 0x33800000.

The root cause is a bug in the __ocl_float2half_rtn function.
Fix is to align with the conformance algorithm, which produces
the correct result (confirmed by AVX2 instruction and SATest Reference).