[[icliscpu-gilr-ics-15959]] Added Xmain documentation PoC
This changeset adds a Xmain-specific section to the LLVM documentation framework
in order to assess this framework as a solution for Xmain's documentation needs
(VPO team will write, update, review and version-control its actual documentation
using this framework as a pilot in the coming few weeks).

Similar to the documents being added here, documents should be added under the
Intel/ folder to clealy mark Xmain's internal documents. Note that since this is
a PoC the project is not configured to build the documents by default.

Old/split llvm commit ID: a12a95f7c55507ab41c4ca7f20c7e64914881265