[LLVM 3.8 UPGRADE] Reflect changes done to user iterarors:
user_begin() now ensures that the module is fully materialized in order
to not miss any uses. On the contrary our BuiltInFuncImport is designed to
work in the conditions when the source module is not fully materialized
and the pass does not assume that it is materialized (to support import
from the built-in library lazy way - a performance improvement made in
the past).