Promote variadic vector arguments of printf function.
SPIR 1.2/2.0/V generators do not promote variadic arguments to printf
function which is OK by the spec. Althoug there is a subtle case with
scalar float to double promotion. C99 spec. demands it to be promoted
while OpenCL C 1.2/2.0 do not mention it in spite of that some devices
may have doubles unsupported.
On the other hand CClang does the promotion of the vector types ATM.
CPU BE implementation of the printf function expects the arguments
to be promoted. So, this commit introduces a pass which does the promotion
of the variadic arguments to overcome these discrepancies/uncertanties.
    charN, ucharN, shortN, ushortN -> uintN
    halfN, floatN                  -> doubleN
Notice what the implementation zero extends all interger types because
the sign information is lost forever but this doesn't matter neither
for our printf implementationd nor for OpenCL C spec. which states that
for vector types the conversion specifier plus the lenght modifier must
match the actual argument type otherwise the behaviour is undefined.