This "fixes" bug 31.  Works around might be a better description.
Take this example:

    ; ModuleID = 'CG-97a11f.cpp'
    target datalayout = "e-m:e-i64:64-n32:64"
    target triple = "lpu"

    define i32 @foo(i1 zeroext %v) {
    entry:
      %call = tail call double @x()
      br i1 %v, label %if.end, label %if.end
    if.end:
      ret i32 0
    }

    declare double @x()

The i1 parameter gets spilled.  Currently, in LPURegisterInfo.td, we claim
the I0 register classes can accomodate type i1.  Unfortunately, this is a lie,
since they can really only handle 0 bits.  However, you can't declare a
register class without specifying which type(s) can be allocated to it.
The net result is that i1 vars can be allocated to I0 classes.  That is the
thing that actually tripped the assertion in LPUInstrInfo.cpp - since it
wasn't prepared to deal with storing/loading 0 bit values.

I have "fixed" this by using ld8/st8 for the I0 register class, which avoids
the compile problem.  But, this is only a work-around for the real issue,
which is the I0 register classes shouldn't be described as allowing i1.
At some point, this should be fixed, but since llvm doesn't recognize an i0
type, we have no options at the moment.  Ideally, we would allow i0 to appear
as a type in codegen, but that's more than I want to bite off currently.
Old/split llvm commit ID: d41aad25909b4d4b6bf6c3b16cec5a808b015918