[LLVM 3.8 UPGRADE] Teach BuiltInFuncImport to link functions with byval parameters.
When we have a case when pSrcFunction has a byval parameter it means the following:

Frontend generated a definition with a byval parameter in one module, for instance:

    define double @_Z18convert_uchar8_satDv8_j(<8 x i32>* byval nocapture readonly align 32) {...}

and a declaration in another module like this one:

    declare double @_Z18convert_uchar8_satDv8_j(<8 x i32>)

The module with declaration perform calls according to the decl:

    %call = tail call double @_Z18convert_uchar8_satDv8_j(<8 x i32> %val_x)

But if we import the definition we can no longer call the function like we did above.

llvm-link solves this problem by creating a ConstantExpr bitcast for
function pointer type, for example:

    %call = tail call double bitcast (double (<8 x i32>*)*
    %@_Z18convert_uchar8_satDv8_j to double (<8 x i32>)*)(<8 x i32> %val_x)

Here we basically do the same thing.