Update unit tests on name mangling.
Fixed following bugs:

1. For purposes of substitution, given a CV-qualified type, the base
type is substitutible, and the type with all the K, V, and r qualifiers
plus any vendor extended types in the same order-insensitive set is
substitutible; any type with a subset of those qualifiers is not. That
is, given a type const volatile foo, the fully qualified type or foo may
be substituted, but not volatile foo nor const foo. Also, note that the
grammar above is written with the assumption that vendor extended type
qualifiers will be in the order-sensitive (not CV) set. An appropriate
grammar modification would be necessitated by an order-insensitive
vendor extended type qualifier like const or volatile. (see
http://mentorembedded.github.io/cxx-abi/abi.html#mangling-type for more
details)

2. Private address space is not encoded during mangling, so it's wrong
to expect demangler to recover private address space qualifier.