Significantly rework the LPUInstrInfo.td file.  The goal is to eventually get to one opcode per actual machine instruction.  Since the LPU currently allows either a register or literal for many operands, and that is not a pre-provided feature of the code generator, this adds a new operand type OPERAND_REG_IMM. In addition, there needed to be pattern matching ("SelectRegImm") for this to come together.  The patterns now generally use "RC" (register/channel) and "RCL" (Register/Channel/Literal) abbreviations also used in the simulator descriptions of the operations.
This checkin goes much of the way there, but not completely.  It does pass
the existing tests, but some operations are not described quite right.  (e.g.
MOV operands are described as having a register class for input, when it is
really a register or an integer or floating point literal.)

Most of the changes in files other than LPUInstrInfo.td are to reflect the now
unified opcodes (e.g. no longer need ADD64i for an add with a literal), or
to add the support code for the new pattern.

The changes in LPUInstrInfo are extensive.  There is now an LPUTypeInfo to
capture the properties of classes of operations.  Unfortunately, this is not
quite complete because there should be a description of operands that can be
(for example) 32b reg OR int literal OR fp literal.  (Currently you can only
describe one or the other.)

This edit also adds PREDPROP and PREDMERGE.
Old/split llvm commit ID: b93720a0ad5f922e13fcbc288cafd5c0dd66d948