Fix for upstream r241143 (git 787b9b4e).
r241143 introduced undefs in the IR. (The latest 3.9 release does the
same.) This is not considered a bug, so we'll handle it. Previous to
this commit, an undef was turned into a disconnected LIC. This change
lowers an undef to a copy from a literal 0.
Old/split llvm commit ID: 7ab537fa54a334e96f0bf6a780b2a826466d37c5