[Workaround] Save and restore StructType names in BIImport
This workaround is applied because LLVM Linker indirectly changes the
Type objects in the RTL module, which should remain unchanged between
compilations.

It is a bug in LLVM Linker, which assumes that the module passed as an
argument to linkInModule() will be destroyed and we can safely change
the Type objects. This is not the case if we clone one module into
another - these two modules would share the same Type objects. When the
Linker changes the Type names in one module, it will do so in the other
as well.

Short reproducer:
  llvm::Module* Module1 = readIR(...);
  llvm::Module* Module2 = readIR(...);

  Module2->dump();
  //
  // %struct.foo.0 = type { i32 }
  //
  // define %struct.foo.0* @f() #0 {
  // entry:
  //   ret %struct.foo.0* null
  // }
  //

  ValueToValueMapTy VMap;
  auto CloneModule2 = CloneModule(Module2, VMap);

  Linker ld(*Module1);
  ld.linkInModule(std::move(CloneModule2));

  Module2->dump();
  //
  //  %0 = type { i32 }         <--- typename changed
  //
  //  define %0* @f() #0 {
  //  entry:
  //    ret %0* null
  //  }
  //

Currently this cannot be easily fixed in LLVM, because it is a part of
the Linker internal Type object tracking. It was introduced in:
  7a551b7c6dd012d67ddf27ab8d87c3e8742c5f11
  Author:     Rafael Espindola <rafael.espindola@gmail.com>
  git-svn-id: https://llvm.org/svn/llvm-project/llvm/trunk@222986
              91177308-0d34-0410-b5e6-96231b3b80d8

  Change how we keep track of which types are in the dest module.

  Instead of keeping an explicit set, just drop the names of types we choose
  to map to some other type.

  This has the advantage that the name of the unused will not cause the context
  to rename types on module read.