[SKX] Fix unresolved zero initialization intrinsics in intrinsics.c.
.. by inclusion of intrin header.

The main purpose of intrinsics.c is to compile intrinsics with
'long long' datatypes as C because OpenCL requires long long to be 128
bit, but LLVM X86 target supports only 64 bit.

This avoided in the past, because our copy of the intrin headers now
must be compiled as both C and OpenCL C. This was achieved using macro
defines.