Implement a new CSAExpandInlineAsm pass, which parses and expands inline into functional MIs.
Inline assembly expansion is a special feature of the CSA Target. Since
no other LLVM targets do this, the functionality is implemented as a
CSA-specific pass.

This feature works best when using GCC-style "extended" syntax, since
this style results in virtual registers flowing through the expanded
MIs. Testing so far has focused on ALU-type operations with C variable
and literal operands.

It's currently possible to refer directly to registers that the backend
knows about (e.g., "%r9" or "%ci64_187"), but this is likely to interact
with optimizations in strange ways, and "clobbers" functionality isn't
implemented yet. Referring directly to registers in asm()s should be
avoided if possible.

Directives such as allocating/declaring LICs are not yet supported.

Finally, if the parser fails to parse anything in an "asm()" expression,
it will give up and revert to the old behavior of simply passing the
text through to the assembly printer. The downside to this is that this
text is then opaque to the backend optimizers, so any resulting
operations will be stuck on the SXU. If this happens, you'll see a
message on stderr announcing it.
Old/split llvm commit ID: 8cb240fe22fe53e9a509ef196d8bb318e2b3cf57