Cope with optional rounding mode operands in CSAOptDFPass
With the appearance of optional rounding modes, the number of operands
is no longer fixed for a given generic operation type. Optional operands
may appear at the end of the operand list. To deal with this, we can
reason about operands relative to the front of the operand list, not the
back. (E.g., for a 3-operand instruction: "get the first two operands
after the first one", rather than "get the last two operands.")

In fact, even for a given MachineInstr with an opcode which has a
rounding mode, the rounding mode MachineOperand may or may not exist,
depending on when/where the MachineInst is created. This is reflected by
MI->getNumOperands(). (But not MI->getDesc().getNumOperands().)
Old/split llvm commit ID: 1d3ee0d1a6e20ccac3a87bb0ac301346bafe904e