[[scels75-13480]] VPlan Predicator: The PredicateRecipes of Regions no longer hold BlockPredicateRecipes themselves.
Up until now each Block (no matter if it was region or BB) was having a BlockPredicateRecipe attached to it (used as a container in the case of Regions).
The problem is that a BlockPredicateRecipe is supposed to generate code, which is not the case for Regions. 

This patch will only attach BlockPredicateRecipes to Basic Blocks only. Regions simply point to the predecessor's predicate recipe.

Updated lit tests to reflect the change.

https://ir-codecollab.intel.com/ui#review:id=64705

Old/split llvm commit ID: c0216b12e09efa59ea4d2ed4bab027bc8a36028f