CSAMemopOrdering: implement whole-function codegen
Previously, memop chains were passed around block by block. For every
alias set, a basic block could have only a single chain input and
output. This made implementation simpler, but we have always known that
it resulted in some unnecessary ordering.

For example, a basic block with only loads in it followed by another basic
block with only loads in it would result in the second block's loads
waiting for the first block's loads. Similarly, a single-block loop with
only loads in it would result in unnecessary ordering backedges.

With this change, ordering chains are described precisely even across
blocks. This is done using two levels of MachineSSAUpdaters such that
all of the work of understanding the CFG is left to LLVM. For example:

   for(...)
     ... = *a;
   *a = ...;

Before, the load would depend on any memops in predecessor blocks, which
includes itself. The result after dataflow conversion was a memory
ordering chain loop, unnecessarily ordering the loads across loop
iterations. Now, the ordering input for all loads is described as coming
from strictly before the loop. The dataflow converter understands that
the input is invariant in the loop and generates repeats of the input
signal for the loads, eliminating the dataflow loop.

Note that this change obsoletes 22823be ("-csa-kill-readchains=true"),
so it has been removed.
Old/split llvm commit ID: bcf1f1890ae04eef26a240f9d83753d72184e964