Fix for passing struct to write_pipe.
The problem was that channel transformation pass was not aware of struct
type of channel. The structs are passed to write_pipe by pointer,
so the pass generated alloca, copied pointer to it and passed pointer to
pointer rather than pointer to struct.
Now in case of the type of value passed to write_pipe is pointer then
the pass uses it directly rather than generate alloca.