Disable cl_channels_altera in common clang by default
This is a better fix to disable channels in clang than the fix in
[5e4779a] Disable cl_altera_channels extension in clang_driver

Channels are disabled in common clang by default. They can be enabled
using a list of device supported extensions.

We add cl_altera_channels in clang driver and not in cpu_device.cpp
where all extensions are defined, because cpu_device does not know
whether we compile for FPGA emulation or not.
In clang_driver we have m_pProgDesc->bFpgaEmulator flag for this
purpose.