[UPSTREAM][HEADER] Split opencl-c.h into separate headers for OpenCL 1.2 and 2.0.
Original design of the opencl-c.h header had a problem with memory
overhead when the header compiled into PCH (or clang module), because
opencl-c.h relied on the __OPENCL_C_VERSION__, and declared some of the
built-ins differently for OpenCL 1.2 and 2.0:

  // A lot of common built-ins (~80% of all)
  #if __OPENCL_C_VERSION__ >= 200
    // OpenCL 2.0 specific builit-ins
  #else
    // OpenCL 1.2 specific builit-ins
  #endif

When the header compiled into PCH, only one value of the
__OPENCL_C_VERSION__ can be used. To support both OpenCL C 1.2 and 2.0,
it was necessary to compile the header twice with different values of
the __OPENCL_C_VERSION__ macro. As a result, both of the compiled
headers contained a copy of the 'common' built-ins.

To avoid this duplication, 'opencl-c.h' header have been splitted into
several headers:
  - 'opencl-c-common.h' with built-ins and types shared between
    OpenCL C 1.2 and 2.0
  - 'opencl-c-20.h' with OpenCL C 2.0 specific built-ins and types
  - 'opencl-c-platform-12.h' with OpenCL C 1.2 platform specific
    declarations
  - 'opencl-c-platform-20.h' with OpenCL C 2.0 platform specific
    declarations
  - 'opencl-c-platform.h' with platform specific declarations of
    built-ins and types shared between OpenCL C 1.2 and 2.0

Meta header 'opencl-c.h' was left for compatibility and it #includes all
other headers in a correct order.

opencl-c-12.h doesn't exist, because it would consists of _only_
platform-specific declarations.

Additional changes:

* Make CLK_NULL_RESERVE_ID invalid reserve id.
* Rename PIPE_RESERVE_ID_VALID_BIT to avoid user name space pollution.
* Current implementation reserve_id_t type assumes that it's a pointer
type whose most significant bit is set to one and the rest of the bits
keep the id value.
* This patch increase reserve id size by one bit on 32-bit platforms and
by 33 bits on 64-bit platforms.

TODO: CORC-2338 Commit this patch to llvm.org

Change-Id: I44a8bba26816daf783f05eb753fa08462186341b
Old/split clang commit ID: 6221a2535428cc92454c91684d32ce1b21635dd7