Fix pipe max_packets misalignment with backing store size
Backing store is a buffer for pipe object: it includes pipe
header (same for all pipes) and an area for packets (depends on pipe
depth) with 'max_packets' size.

This 'max_packets' should be aligned b/w allocation and pipe_init and
previously it was used in 2 different places:
  - in pipe_get_total_size() used to compute a size of backing store
  - in pipe_init() function

Moved this into new function called __pipe_get_max_packets(): for pipe
depth it computes a 'max_packets' size we need, and it is used for
both backing store allocation and pipe_init.

Renamed 'max_packets' argument in pipe_init() into 'depth', because it
is really a depth of a pipe - max_packets may be higher than this.

Change-Id: I59bdbc706bac794cf75344d8a46be27a32b39fc1