[LLVM UPGRADE 4.0] Added debug info to call inst which calls func with debug info.
The following check was added in LLVM4.0:
"inlinable function call in a function with debug ".
In PrepareKernelArgs pass we create a wrapper function which calls the
original function. The original function will be inlined later. But when
original function has debug info we need to attach "dummy" debug info to
call inst to satisfy the check above.