CSA: Have Gnu ToolChain's GCCInstallationDetector treat CSA like x86_64
The main effect of this is that it allows Clang to find candidate
GCC/libstdc++ installations for CSA : the x86_64 ones. This is what was
already happening before this latest Clang pulldown, but I guess the
logic was moved.

This is surely not always the right thing to do, but it allows us to at
least compile some C++ for now.
Old/split clang commit ID: 55902c13c7784b114371c1638149ec0e0eca2c0e