Adding implementation of pragma inline.
This is an intel pragma that applies to the next compound statement.
The attribute at a call site, indicates the inline state of the function
called. This attribute can take an option, recursive that indicates that
the callee are also to be inlined.

Change-Id: I1e82870f2232dc5509fc2ed94346d8e45b63887b
Old/split clang commit ID: f8fb1d938efe425921cd2b934b56410e157b41e2