Fix substitution rules for mangler/demangler
According to Itanium ABI we should treat as substitution candidates:
* base type if it is substitutable,
* qualified type with all qualifiers including address spaces,
* pointer type if it is a pointer

It means that for "const clk_event_t *" type we can substitute:
* clk_event_t | base type
* const clk_event_t | qualified type
* const clk_event_t * | pointer type

Previously, we didn't treat "const clk_event_t" as substitutable entity,
but substituted "clk_event_t *" which is not according to Itanium ABI.

This patch introduce a workaround to fix the issue by adding to list of
substitutions a dummy value which represents "qualified" type. In current
mangler/demangler design we cant properly handle the issue due to only
PointerType keeps qualifiers and we can't create any other qualified type.