CSAMemopOrdering: avoid redundant PHIs; connect non-section memops in a region
- Avoid generating redundant PHIs by directly using the edge flowing out
of the section as a region mergee. (Previously, the edge's memop's output
was used.)
- Avoid leaving the graph disconnected: previously the edge flowing into
the section had its destination redirected, but the source was left
unmodified. This is fine, assuming that all memops in a region are in a
section. However, in the case where there were memops in the region but
NOT in a section, this can leave them disconnected. Now the incoming
edge source is included in the region merge.
Old/split llvm commit ID: 0860a8dd51616f707c5001baa05bb91d45e138a5