[OpenCL] Add cl_intel_vec_len_hint extension support.
This extension adds kernel attribute "__attribute__((intel_vec_len_hint(<int>)))".
Which can be used to provide a hint to the compiler that the kernel performs
the best if vectorized to a specified vector length.

The lengths accepted by the attribute are:
0 - the compiler makes heuristic–based decision whether to
        vectorize a kernel, and if so which vector length to use (default behavior).
1 - No vectorization by the compiler. Explicit vector data
        types in kernels are left intact.
4 - Disables heuristic and vectorizes to the length of 4.
8 - Disables heuristic and vectorizes to the length of 8.

Other values are invalid. An error will be reported during
compilation.

If you specify CL_CONFIG_CPU_VECTORIZER_MODE config variable it overrides
vectorization width set through the attribute.
Note! If CL_CONFIG_CPU_VECTORIZER_MODE is set to 0 it forces autovectorization
even if attribute is set to specific vectorization width.