[InstCombine] Add a few additional tests for the icmp size reduction optimization.
This includes a test for a case we currently aren't handling because we missed a commuted check on some of the binops.

There are also TODOs for cases we shouldn't have been optimizing when a subtract is involved.

Change-Id: I8d1d7eff5676d8adb16b3f8d587f1dc857573c7d
Old/split llvm commit ID: b58b0c3db040da7f33a76917a9491fc1c03fb8fe