Improved work-group fusion in case of one thread.
The optimization is to make local size = global size if we have local
execution context(the case when we have queue with master thread only)
and nullptr is passed to clEnqueueNDRangeKernel as local size.

It doesn't make sense to have NDRange split to several WGs if we are
executing in one thread, we only paying for WGs dispatching.