[InstCombine] Rewrite icmp size reduction to base its decisions on DataLayout.
Instead of assuming that 8/16/32 are legal types, look for the patterns we recognize and then query data layout to determine if the bit width we found is legal.

This also fixes the shl/binop/shr pattern to support the shl being on either operand of commutable operations, and only on the RHS of sub. This matches the combine in InstCombiner::FoldShiftByConstant that this is based on.

Change-Id: Ib0c5fc1ba38f0eb47d4bf5eddb0dc3dd43c6f030
Old/split llvm commit ID: 4fde78830051f128665164745fea8ff2e550405b