old ("wavefront") memop ordering: rip out a problematic optimization
This optimization was one of a slew intended to improve compile time,
but it was rarely missing section-ordering edges due to a bug. This
resulted in slower run time. (This was seen in a SAP kernel, which we
can't have sources for, unfortunately.)

With the new default/"independent" memop ordering this doesn't matter
too much anymore, but we might as well fix code that's checked in.
Old/split llvm commit ID: 516b66087d1cc14064bfe58350efe1af19962aea