Fix for CMPLRS-46103
1) Do not consider i1 types during VF candidate selection.

2) Use the type on the load/store Expr instead of trying to derive the type from the pointer type.

Example: array of struct types may be represented as [1024 x { float, float }], but the load
expr of an element of the struct will already reflect that it is a float load. This saves from
having to start with [1024 x { float, float }] and trying to derive that we're loading a float
from it. Also, it would not be possible for the existing code that attempts this derivation to
be able to determine the type of load for a struct like { float, int }.

Change-Id: I4f9f8ad32ce8f7b9ac4e8a5d734e83c660b03951
Old/split llvm commit ID: 22861e1edcf9e3a7f96af92116877cd7117ee81d