Implement a basic streaming memory converter pass.
The implementation in CSADataflowSimplify.cpp is a bit of a starter framework
for developing simple conversion passes in the style of InstCombiner. As Pablo
is implementing a matching framework right now, I'm being rather conservative
in putting functionality here.

In the longer run, it seems clear to me that we'll need some sort of better
dataflow structure for recognizing the sources and dependencies of operands;
the current tracking abilities that derive from the MachineInstr level are just
too weak to do this without layering something on top. Another key issue that
seems problematic is the current layering interface for differentiating between
SXU and non-SXU allocation of the instructions (the MachineInstr flag is rather
hidden).
Old/split llvm commit ID: ef4bc69880e2e9a81dfbc8832f322a78d8d22218