Add a pass to push switches of alls to alls of switches.
This helps for conversion of a streaming memcpy implementation: the memorder
lowering pass lowers the loops into all'ing the memory orders for each
iteration and then switching on the loop iteration stream. With this mini pass,
the order is reversed, so the streaming load can detect that no one cares about
the switches.

The memcpy loop is still not quite properly generated, but it's much closer.
Old/split llvm commit ID: e9bf61d4da50da208929c23cea07394793977570