Fix PipeSupport to avoid a deadlock in pipes internal caching
This patch fixes a deadlock, which can occur because of pipe built-ins
internal caching.

When data packets are pushed to the pipe, they are not immediately
visible by the consumer: they are cached until more packets arrive,
and when a certain limit is reached, all cached packets become
visible. This approach helps to reduce a synchronization overhead, but
it also causes problems, when reader expects that all data will be
transmitted immediately (like in GF RBE 1.2):

*GF RBE 1.2 issue*
-----------------

when 2 kernels are communicating over channels:
        ..
------  -----> ------
| A  |         | B  |
------  -----> ------
        xxxxx

- Kernel A writes 2 packets, but they are cached, i.e. kernel B
  doesn't see them yet.
- Then Kernel A writes 5 packets (2 of them may be cachec, but that's
  irrelevant)
- According to the algorithm, kernel B reads packets from 1st pipe,
  then reads frin 2nd pipe

Since packets in the 1st pipe are cached, kernel B cannot read
anything, and kernel A is blocked on 2nd pipe ('pipe full' condition).

*Patch changes*
---------------

To solve this problem, the patch adds the following:
  - when we are in a state of a possible deadlock, we must flush
    everything
  - a deadlock may occur only if we are stuck on a pipe read/write
    call

For every function we look for the pipe calls, and add a conditional:
if a pipe call failed (-1 is returned) - call a __flush_all function.

Note that only non-blocking pipes are handled: the pass expects that
blocking pipes are already resolved.

*Flush function*
---------------

__flush_all function is created by the PipeSupport pass for *all*
functions, which have pipe calls. For example, for 2 functions 'foo'
and 'bar', which both use pipes, we would have '__flush_all_foo' and
'__flush_all_bar' helper functions.

All pipes used by a function are passed to a corresponding
__flush_all() functions as parameters: global pipes are loaded at
caller side, local pipes are passed as is.

Change-Id: I89fb7db1f43fc1ae2ae44ae1a2529d672dddcc55