Resolve blocking pipe built-ins in PipeSupport
PipeSupport relies on 2 facts:
  - blocking built-ins are implemented as a loop over non-blocking
    built-ins

  - they are inlined, so PipeSupport can modify this loop based on a
    call site

Previously blocking pipe built-ins were implemented in RTL (with
always_inline attribute), and in order to inline them we had to:
  1. run BIImport (import bodies for inlining)
  2. run Inliner
  3. run PipeSupport, which transformes inlined built-ins
  4. run BIImport again, because PipeSupport could insert a new
     built-in

This patch eliminates steps 1 and 2.

Instead of having separate function for blocking pipe built-ins are
running BIImport and Inliner to resolve them, this patch resolves
these built-ins by hand.

It replaces a blocking __read_pipe_2_bl() built-in with a spinlock
around a non-blocking __read_pipe_2() built-in.

Change-Id: Ie614839f68be95e80ea8880932947184224b3a23