Use generic opcodes to simplify a lot of the CSAInstrInfo helper methods.
At this point, the commit diff of this change series removes more lines than
it adds--even including the entire tablegen backend. There's still more code
to be removed. Some of the compare logic is untouched because the tablegen
is producing bad information (CMPO*F and CMPU*F end up mapping to the same
generic opcodes).

The logic of how to handle some of the issues around requesting opcodes that
don't actually exist is somewhat problematic. The current interface of
adjustOpcode is cumbersome, but based on the logic of the functions it's
simplifying, so facts on the ground seem to rule. By contrast, the decision
of bool-did-it-work-plus-outparam or optional-outparam-for-it-worked is even
worse, since many cases plainly don't care (it should always exist), and I
don't want to explode function signatures for instances that don't really
come up.
Old/split llvm commit ID: 8e4e0cb23f20a03ea5098038fc09d9695a9d7132