Add tests from backend/tests/opencl to automatic testing
One can launch them with ics libbuild check-ocl-backend-opencl -j16
command.

The idea was to enable a vec_len_hint unit test in ics automatic testing
system. Many of the tests there started to fail so they were marked as XFAIL
and have to be fixed later. The issue that was created:
https://jira01.devtools.intel.com/browse/CORC-2413

Change-Id: I5b252f3b5d7f048d7bf9b040fd89e1f14d79225d